+++ wxc_sdk/all_types.py
@@ -23,7 +23,7 @@ from wxc_sdk.common.schedules import Event, RecurWeekly, RecurYearlyByDate, Recu
 from wxc_sdk.device_configurations import DeviceConfiguration, DeviceConfigurationOperation,\
     DeviceConfigurationResponse, DeviceConfigurationSource, DeviceConfigurationSourceEditability,\
     DeviceConfigurationSources
-from wxc_sdk.devices import ActivationCodeResponse, Device, ProductType, TagOp
+from wxc_sdk.devices import ActivationCodeResponse, ConnectionStatus, Device, ProductType, TagOp
 from wxc_sdk.events import ComplianceEvent, EventData, EventResource, EventType
 from wxc_sdk.groups import Group, GroupMember
 from wxc_sdk.licenses import License, SiteType
@@ -172,16 +172,16 @@ __all__ = ['AcdCustomization', 'Action', 'ActivationCodeResponse', 'AdaptiveCard
            'CallerIdSelectedType', 'CallingBehavior', 'CallingCDR', 'CallingLineId', 'CallingPermissions',
            'CallingType', 'CallsFrom', 'CapabilityMap', 'ChatObject', 'ClosedCaption', 'CnameRecord', 'CoHost',
            'CodeAndReason', 'ComfortMessageBypass', 'ComfortMessageSetting', 'CommonDeviceCustomization',
-           'ComplianceEvent', 'CreateInviteesItem', 'CreateLocationFloorBody', 'CreateMeetingBody',
-           'CreateMeetingInviteeBody', 'CreateMeetingInviteesBody', 'CreateResponse', 'CustomNumbers', 'Customer',
-           'CustomizedQuestionForCreateMeeting', 'DND', 'DectCustomization', 'DectDevice', 'DefaultAudioType',
-           'DefaultVoicemailPinRules', 'DeleteTranscriptBody', 'DestinationType', 'Device', 'DeviceActivationState',
-           'DeviceConfiguration', 'DeviceConfigurationOperation', 'DeviceConfigurationResponse',
-           'DeviceConfigurationSource', 'DeviceConfigurationSourceEditability', 'DeviceConfigurationSources',
-           'DeviceCustomization', 'DeviceCustomizations', 'DeviceManagedBy', 'DeviceManufacturer', 'DeviceMember',
-           'DeviceMembersResponse', 'DeviceOwner', 'DeviceStatus', 'DeviceType', 'DialPatternStatus',
-           'DialPatternValidate', 'DialPatternValidationResult', 'DialPlan', 'DialResponse', 'Dialing',
-           'DirectoryMethod', 'DisplayCallqueueAgentSoftkey', 'DisplayNameSelection', 'DistinctiveRing',
+           'ComplianceEvent', 'ConnectionStatus', 'CreateInviteesItem', 'CreateLocationFloorBody',
+           'CreateMeetingBody', 'CreateMeetingInviteeBody', 'CreateMeetingInviteesBody', 'CreateResponse',
+           'CustomNumbers', 'Customer', 'CustomizedQuestionForCreateMeeting', 'DND', 'DectCustomization',
+           'DectDevice', 'DefaultAudioType', 'DefaultVoicemailPinRules', 'DeleteTranscriptBody', 'DestinationType',
+           'Device', 'DeviceActivationState', 'DeviceConfiguration', 'DeviceConfigurationOperation',
+           'DeviceConfigurationResponse', 'DeviceConfigurationSource', 'DeviceConfigurationSourceEditability',
+           'DeviceConfigurationSources', 'DeviceCustomization', 'DeviceCustomizations', 'DeviceManagedBy',
+           'DeviceManufacturer', 'DeviceMember', 'DeviceMembersResponse', 'DeviceOwner', 'DeviceStatus', 'DeviceType',
+           'DialPatternStatus', 'DialPatternValidate', 'DialPatternValidationResult', 'DialPlan', 'DialResponse',
+           'Dialing', 'DirectoryMethod', 'DisplayCallqueueAgentSoftkey', 'DisplayNameSelection', 'DistinctiveRing',
            'EmergencyDestination', 'EnabledAndNumberOfDays', 'EnabledAndValue', 'EntryAndExitTone',
            'ErrorMessageObject', 'ErrorObject', 'Event', 'EventData', 'EventResource', 'EventType',
            'ExecAssistantType', 'ExpirePasscode', 'ExternalCallerIdNamePolicy', 'ExternalTransfer', 'FailedAttempts',